requirements.txt
pyserial 
nbconvert
file_read_backwards
pytest